REFACTOR: complete change of DAOs, DAO segregate. Comments and proposals what more we can improve
