+++ src/main/java/dao/IAdminDAO.java
@@ -9,9 +9,11 @@ import java.sql.SQLException;
 import java.util.List;
 
 public interface IAdminDAO{
-    
 
-    List<Codecooler> getCodecoolers(int roomId) throws DBException;
+    /*this interface contains methods which are left after splitting everything into other interfaces -
+    this are two admin's methods
+    both are still NOT implemented in AdminDAO*/
+
     void addLevel();
     void createNewClass();
 }
