+++ src/main/java/dao/ICodecoolerDAO.java
@@ -1,24 +1,16 @@
 package  dao;
 
-
 import java.sql.SQLException;
-import model.items.Artifact;
 import model.users.Codecooler;
 import model.users.User;
-
 import java.util.List;
 
 public interface ICodecoolerDAO {
+    //this interface contains methods to create and update codecooler
+    // and show his level
 
     void createCodecooler(User user, Codecooler codecooler);
-    int showWallet(int id) throws SQLException;
-
-    void buyArtifact(int userID, int artifactID) throws SQLException;
-
-    List<Artifact> showBoughtArtifacts(int id) throws SQLException;
-
-    void buyArtifactWithTeammates();
-
+    List<Codecooler> getCodecoolersFromRoom(int roomId) throws DBException; //what about this exception?
     int showLevelOfExperience(int id) throws SQLException;
 }
